Изменить int на long в некоторых местах (#92)
* 88: replace int with long only in db and model

* 88: replace int with long in other code

* 88: add tests

+++ server/src/main/java/ru/spbstu/tesseract/dto/AssetLongDto.java
@@ -11,8 +11,8 @@ public class AssetLongDto {
     private int assetId;
     private String assetTitle;
     private String assetDescription;
-    private int assetPrice;
-    private int assetPriceDiff;
+    private long assetPrice;
+    private long assetPriceDiff;
     private String companyTitle;
     private String companyDescription;
     private int riskTypeId;
